pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aorta4llm"
version = "0.2.2"
description = "Organizational reasoning for LLM agent systems"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
aorta = "aorta4llm.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aorta4llm"]

[tool.hatch.build.targets.wheel.force-include]
"src/aorta4llm/org-specs/templates" = "aorta4llm/cli/org-specs/templates"

[tool.pytest.ini_options]
testpaths = ["tests"]
